Update and a lot of new changes, hope to redeem myself.

+++ mint_i3wm/compton.sh
@@ -1,61 +1,52 @@
-#!/bin/bash
+#!/bin/sh
 
-if [ compton ]
-then
-    sudo apt remove compton compton-conf -y
-fi
+[ "$1" ] && flag="$1" || flag='-i'
 
-if [ xsltproc ]
-then
-    sudo apt remove xsltproc -y
-fi
+way=$(pwd)
 
-dir=~/mint_i3wm/
+[ $(which compton) ] && sudo apt remove compton compton-conf -y
 
-if [ $1 == '-i' ] 2> /dev/null
-then
-
-	sudo apt install libx11-dev xsltproc libxml2-dev docbook-xsl libgl1-mesa-dev libxcomposite-dev libxdamage-dev libxfixes-dev libxext-dev libxrender-dev libxrandr-dev libxinerama-dev pkg-config make x11proto-core-dev x11-utils libpcre3 libpcre3-dev libconfig-dev libdrm-dev libgl1-mesa-glx libdbus-1-dev asciidoc -y 
-
-	git clone https://github.com/tryone144/compton.git && cd $dir/compton
+compileCompton() {
+	git clone https://github.com/tryone144/compton.git && cd compton
 
 	sudo make
 	sudo make install
+}
 
-	if [ ! -f ~/.config/compton.conf ]
-	then
-    	>> ~/.config/compton.conf
-	fi
-
-	cp ~/.config/compton.conf ~/.config/compton-conf.bak
-	cp $dir/comptoncfg ~/.config/compton.conf
-
-elif [ $1 == '-u' ] 2> /dev/null
+if [ "$flag" = '-i' ] 2> /dev/null
 then
 
-    if [ ! xlstproc ]
-    then
-        sudo apt install xlstproc libxml2-dev docbook-xsl -y
-    fi
-	
-	git clone https://github.com/tryone144/compton.git && cd $dir/compton
+	sudo apt install libx11-dev libxml2-dev docbook-xsl \
+	 libgl1-mesa-dev libxcomposite-dev libxdamage-dev libxfixes-dev \
+	 libxext-dev libxrender-dev libxrandr-dev libxinerama-dev pkg-config \
+	 make x11proto-core-dev x11-utils libpcre3 libpcre3-dev libconfig-dev \
+	 libdrm-dev libgl1-mesa-glx libdbus-1-dev asciidoc -y 
 
-	sudo make
-	sudo make install
-	sudo make uninstall
+	cd /tmp
+	compileCompton
+	cd $way
 
-	if [ -f ~/.config/compton.conf ]
-	then
-		rm ~/.config/compton.conf
-	fi
+	[ ! -f ~/.config/compton.conf ] && > ~/.config/compton.conf
 
-    if [ -f ~/.config/compton-conf.bak ]
-    then
-        mv ~/.config/compton-conf.bak ~/.config/compton.conf
-    fi
-    
+	cp -v ~/.config/compton.conf ~/.config/compton-conf.bak
+	cp -v $way/comptoncfg ~/.config/compton.conf
+
+elif [ "$flag" = '-u' ] 2> /dev/null
+then
+
+   cd /tmp
+   cd compton/ 2> /dev/null || compileCompton 
+   sudo make uninstall
+   cd $way
+
+   [ -f ~/.config/compton.conf ] && rm -v ~/.config/compton.conf
+
+   [ -f ~/.config/compton-conf.bak ] &&
+    mv -v ~/.config/compton-conf.bak ~/.config/compton.conf
+
+   rm -rfv /tmp/compton
     sudo apt autoremove -y
         
 else
-	echo "${highlight}Don't forget to use a flag (-i or -u).${reset}"
+	echo "${highlight}Invalid flag, please use a valid one (-i or -u).${reset}"
 fi
